Move namespaces and directories and normalize composer.json
 - Typically the package namespace should match the repository name.
 - Instead of sub-folders, specify using psr-4 root namespace.
 - Explicity require packages our code uses rather than allowing
   them to be pulled in as third-party indirect dependencies.
 - Unit tests are typically under "tests" folder.
 - Shorten class names since namespace clarifies type of library.
 - Normalized composer.json using `ergebnis/composer-normalize`.
