+++ composer.json
@@ -14,28 +14,31 @@
   ],
   "require": {
     "php": ">=7.4",
+    "doctrine/orm": "^2.6",
+    "illuminate/contracts": "^8.0",
     "illuminate/queue": "^8.0",
+    "illuminate/support": "^8.0",
     "laravel-doctrine/orm": "^1.5"
   },
   "require-dev": {
-    "phpunit/phpunit": "^9.3",
-    "mockery/mockery": "^1.3.1"
-  },
-  "autoload": {
-    "psr-4": {
-      "Pinnacle\\Queue\\": "src/Pinnacle/Queue"
-    }
-  },
-  "autoload-dev": {
-    "psr-4": {
-      "test\\Pinnacle\\Queue\\": "test"
-    }
+    "mockery/mockery": "^1.3.1",
+    "phpunit/phpunit": "^9.3"
   },
   "extra": {
     "laravel": {
       "providers": [
-        "Pinnacle\\Queue\\DoctrineQueueWorkerServiceProvider"
+        "Pinnacle\\DoctrineQueueWorker\\WorkerServiceProvider"
       ]
+    }
+  },
+  "autoload": {
+    "psr-4": {
+      "Pinnacle\\DoctrineQueueWorker\\": "src/"
+    }
+  },
+  "autoload-dev": {
+    "psr-4": {
+      "Pinnacle\\DoctrineQueueWorker\\Tests\\": "tests"
     }
   }
 }
